split formatting tests
